# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LANDSAFE - Land Stability and Safety Framework API
security:
  - bearerAuth: []
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      security: []
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change current user's password and revoke existing tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
        "400":
          description: Invalid new password
        "401":
          description: Current password incorrect or unauthorized
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /alerts:
    get:
      operationId: getAlerts
      tags: [alerts]
      summary: List alerts
      parameters:
        - name: locationId
          in: query
          schema:
            type: integer
        - name: priority
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, acknowledged, resolved]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Alerts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}:
    get:
      operationId: getAlert
      tags: [alerts]
      summary: Get alert by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
    patch:
      operationId: updateAlert
      tags: [alerts]
      summary: Update alert status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertUpdate"
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /devices:
    get:
      operationId: getDevices
      tags: [devices]
      summary: List devices
      responses:
        "200":
          description: Devices list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      operationId: createDevice
      tags: [devices]
      summary: Create device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get device by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
    patch:
      operationId: updateDevice
      tags: [devices]
      summary: Update device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceUpdate"
      responses:
        "200":
          description: Updated device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
    delete:
      operationId: deleteDevice
      tags: [devices]
      summary: Delete device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /ai/risk-analysis/{locationId}:
    get:
      operationId: getRiskAnalysis
      tags: [ai]
      summary: Get AI risk analysis for a location
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risk analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskAnalysis"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sensorToken:
      type: apiKey
      in: header
      name: x-sensor-token
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        rememberMe:
          type: boolean
    ChangePasswordInput:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 12
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [public, operator, admin]
        isActive:
          type: boolean
        createdAt:
          type: string
        lastLogin:
          type: ["string", "null"]
...
    Alert:
      type: object
      required: [id, locationId, alertType, priority, status, message, createdAt]
      properties:
        id:
          type: integer
        locationId:
          type: integer
        locationName:
          type: ["string", "null"]
        alertType:
          type: string
          enum: [threshold_exceeded, rainfall_warning, sensor_offline, battery_low, ground_movement, earthquake, emergency_notification]
        priority:
          type: string
          enum: [critical, high, medium, low]
        status:
          type: string
          enum: [active, acknowledged, resolved]
        message:
          type: string
        operatorId:
          type: ["integer", "null"]
        operatorName:
          type: ["string", "null"]
        acknowledgedAt:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    AlertUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [acknowledged, resolved]
        operatorNote:
          type: string
    Device:
      type: object
      required: [id, deviceId, locationId, status, batteryLevel, signalStrength, lastCommunication, createdAt]
      properties:
        id:
          type: integer
        deviceId:
          type: string
        locationId:
          type: integer
        locationName:
          type: ["string", "null"]
        firmwareVersion:
          type: string
        status:
          type: string
          enum: [online, offline, maintenance, calibration]
        batteryLevel:
          type: number
        signalStrength:
          type: number
        lastCommunication:
          type: string
        installationDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    DeviceInput:
      type: object
      required: [deviceId, locationId]
      properties:
        deviceId:
          type: string
        locationId:
          type: integer
        firmwareVersion:
          type: string
        installationDate:
          type: string
        notes:
          type: string
    DeviceUpdate:
      type: object
      properties:
        firmwareVersion:
          type: string
        status:
          type: string
          enum: [online, offline, maintenance, calibration]
        notes:
          type: string
    RiskAnalysis:
      type: object
      required: [locationId, locationName, riskLevel, probability, confidence, analyzedAt]
      properties:
        locationId:
          type: integer
        locationName:
          type: string
        riskLevel:
          type: string
          enum: [normal, watch, alert, danger]
        probability:
          type: number
        confidence:
          type: number
        potentialTriggers:
          type: array
          items:
            type: string
        historicalSimilarEvents:
          type: integer
        recommendedActions:
          type: array
          items:
            type: string
        evacuationSuggested:
          type: boolean
        explanation:
          type: string
        analyzedAt:
          type: string
...